remove ConfigParser requirement
This removes the ConfigParser import from the module, and also removes
the ConfigParser requirement noted in the documentation.  There
remains a reference to ConfigParser in the "notes" section, concerning
the behavior w/r/t a "default" section; I haven't tested the behavior
of the module in that case so I have left the reference in place.

Closes #1.